Rebuild of common.css stylesheet
Changed to a 800 pixel fixed width layout. Based on 'Elastic Constrained' layout in 'Pro CSS Techniques', which is laid out in ems calculated on a reduction of standard 16px font size to 10 (62.5%) so 1 em = 10px. Floated content and navigation sections are then contrained by percentages within an overall content area measured in ems.